[리팩토링] 이체 금액 자료형 변경
- type: int -> long

+++ backend/src/main/java/com/shinhan/walfi/dto/transfer/LocalTransferDTO.java
@@ -12,7 +12,7 @@ public class LocalTransferDTO {
 
     private String 입금은행코드;
 
-    private int 이체금액;
+    private long 이체금액;
 
     private String 입금계좌번호;
